+++ partials/entry/readmore.php
@@ -10,20 +10,22 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+// Define read more icon.
 $icon = '';
-if ( is_rtl() ) {
-	$icon = oceanwp_icon( 'angle_left', false );
-} else {
-	$icon = oceanwp_icon( 'angle_right', false );
-}
+$icon = is_rtl() ? oceanwp_icon( 'angle_left', false ) : oceanwp_icon( 'angle_right', false );
 
+$blog_continue_reading_content = '';
+ob_start();
+?>
+<a href="<?php the_permalink(); ?>"><?php oceanwp_theme_strings( 'owp-string-post-continue-reading' ); ?><span class="screen-reader-text"><?php the_title(); ?></span><?php echo $icon; ?></a>
+<?php
+$blog_continue_reading_content .= ob_get_clean();
 ?>
 
 <?php do_action( 'ocean_before_blog_entry_readmore' ); ?>
 
 <div class="blog-entry-readmore clr">
-	<a href="<?php the_permalink(); ?>" title="<?php oceanwp_theme_strings( 'owp-string-post-continue-reading', 'oceanwp' ); ?>"><?php oceanwp_theme_strings( 'owp-string-post-continue-reading', 'oceanwp' ); ?><?php echo $icon; ?></a>
-	<span class="screen-reader-text"><?php the_title(); ?></span>
+	<?php echo $blog_continue_reading_content; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
 </div><!-- .blog-entry-readmore -->
 
 <?php do_action( 'ocean_after_blog_entry_readmore' ); ?>